[C++20 Modernization -> core/project settings files] Split files, added explanations of what is happening, improved memory and speed where possible.
