+++ core/config/project_settings_signals.cpp
@@ -1,0 +1,66 @@
+/**************************************************************************/
+/*  project_settings_signals.cpp                                          */
+/**************************************************************************/
+/*                         This file is part of:                          */
+/*                             REDOT ENGINE                               */
+/*                        https://redotengine.org                         */
+/**************************************************************************/
+/* Copyright (c) 2024-present Redot Engine contributors                   */
+/*                                          (see REDOT_AUTHORS.md)        */
+/* Copyright (c) 2014-present Godot Engine contributors (see AUTHORS.md). */
+/* Copyright (c) 2007-2014 Juan Linietsky, Ariel Manzur.                  */
+/*                                                                        */
+/* Permission is hereby granted, free of charge, to any person obtaining  */
+/* a copy of this software and associated documentation files (the        */
+/* "Software"), to deal in the Software without restriction, including    */
+/* without limitation the rights to use, copy, modify, merge, publish,    */
+/* distribute, sublicense, and/or sell copies of the Software, and to     */
+/* permit persons to whom the Software is furnished to do so, subject to  */
+/* the following conditions:                                              */
+/*                                                                        */
+/* The above copyright notice and this permission notice shall be         */
+/* included in all copies or substantial portions of the Software.        */
+/*                                                                        */
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,        */
+/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF     */
+/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. */
+/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY   */
+/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,   */
+/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE      */
+/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
+/**************************************************************************/
+
+#include "project_settings.h"
+
+/*
+ * Marks project settings as changed, schedules notification to be
+ * emitted later.
+ */
+void ProjectSettings::_queue_changed() {
+	CallQueue *msg_queue = MessageQueue::get_singleton();
+
+	const bool already_queued = is_changed;
+	const bool queue_unavailable = (msg_queue == nullptr);
+	const bool buffer_empty = (msg_queue && msg_queue->get_max_buffer_usage() == 0);
+
+	if (already_queued || queue_unavailable || buffer_empty) {
+		return;
+	}
+
+	is_changed = true;
+
+	callable_mp(this, &ProjectSettings::_emit_changed).call_deferred();
+}
+
+/*
+ * Emit `settings_changed_` if queued.
+ */
+void ProjectSettings::_emit_changed() {
+	if (!is_changed) {
+		return;
+	}
+
+	is_changed = false;
+
+	emit_signal("settings_changed");
+}
